+ model back to int! + małe zmiany w fizyce
git-svn-id: svn+ssh://gamma/home/samba/koturj/svn/inz@112 e70bc5d4-e669-4a26-a90f-bfde2a246bde

+++ src/mem/db/planet_row.h
@@ -27,7 +27,7 @@ namespace MEM
 			float xvel;
 			float yvel;
 			float zvel;
-			float model_id;
+			int model_id;
 
 		private:
 			static const std::string save_string;
